+++ NEWS.md
@@ -9,6 +9,9 @@
     
 #### MINOR IMPROVEMENTS
 * The package no longer depends on `EBImage` package from Bioconductor. `EBImage` is a great package but CRAN packages are easier to install if they depend on CRAN packages only.
+
+#### GONE
+* `can_be_integer()` is gone. Use `checkmate::check_integerish()` instead.
     
 
 ## 1.0.0